changed help and subcommand layout
